revise API query logic
Querying the API is now a two-step process: first make a search query
with the given string and return a list of results. Second, index into
the first result to get the food id and make a food id query to get
detailed info on that food. This allows us to access the RACC values.